+++ pisces/schema/css3.py
@@ -455,7 +455,7 @@ szz = Column(Float(24),
              info={'default': -100000000, 'parse': parse_float, 'dtype': 'float', 'width': 15, 'format': '15.4f'})
 
 tagid = Column(Integer,
-               info={'default': -1, 'parse': parse_float, 'dtype': 'float', 'width': 8, 'format': '8d'})
+               info={'default': -1, 'parse': parse_int, 'dtype': 'int', 'width': 8, 'format': '8d'})
 
 tagname = Column(String(8),
                  info={'default': '-', 'parse': parse_str, 'dtype': 'a8', 'width': 8, 'format': '8.8s'})
@@ -931,7 +931,7 @@ class Wftag(Base):
 
     @declared_attr
     def __table_args__(cls):
-        return (PrimaryKeyConstraint('tagid'), UniqueConstraint('tagname', 'tagid', 'wfid'),)
+        return (PrimaryKeyConstraint('tagname', 'tagid', 'wfid'),)
 
     tagname = tagname.copy()
     tagid = tagid.copy()